pollsproject/requirements.txt
django==5.0.2
django-ninja==1.1.0
psycopg==3.1.18
psycopg-binary== 3.1.18
python-dotenv==1.0.1